backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
python-multipart==0.0.9
python-dotenv==1.0.1
SQLAlchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openai==1.30.1
pdfplumber==0.11.0
pypdf==4.2.0
python-docx==1.1.2
aiofiles==23.2.1
